fix(tui): improve drag selection and input click behavior
